# [perf] Add -O3 to verilation and remove optimization from compilation